first part of splitting the integration tests
