+++ .github/workflows/pull-request.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Install license-header-checker
         run: curl -s https://raw.githubusercontent.com/lluissm/license-header-checker/master/install.sh | bash
       - name: Run license check
-        run: ./bin/license-header-checker -a -r .github/license-header.txt . py && [[ -z `git status -s` ]]      
+        run: ./bin/license-header-checker -a -r .github/license-header.txt . py && [[ -z `git status -s` ]]
 
 #      - name: Lint
 #        run: |
