+++ tests/integration-tests/client_test.py
@@ -17,108 +17,12 @@
 # under the License.
 #
 
-import random
 import string
-import time
 
 import pytest
 
 from xata.client import XataClient
 from xata.errors import BadRequestException, RecordNotFoundException
-
-
-def create_demo_db(client, db_name):
-    client.put(f"/dbs/{db_name}", cp=True, json={"region": "us-east-1"})
-
-    client.put(f"/db/{db_name}:main/tables/Posts")
-    client.put(f"/db/{db_name}:main/tables/Users")
-    client.put(
-        f"/db/{db_name}:main/tables/Posts/schema",
-        json={
-            "columns": [
-                {"name": "title", "type": "string"},
-                {"name": "labels", "type": "multiple"},
-                {"name": "slug", "type": "string"},
-                {"name": "text", "type": "text"},
-                {
-                    "name": "author",
-                    "type": "link",
-                    "link": {
-                        "table": "Users",
-                    },
-                },
-                {"name": "createdAt", "type": "datetime"},
-                {"name": "views", "type": "int"},
-            ]
-        },
-    )
-
-    client.put(
-        f"/db/{db_name}:main/tables/Users/schema",
-        json={
-            "columns": [
-                {"name": "name", "type": "string"},
-                {"name": "email", "type": "email"},
-                {"name": "bio", "type": "text"},
-            ]
-        },
-    )
-
-
-def delete_db(client, db_name):
-    client.delete(f"/dbs/{db_name}", cp=True)
-
-
-def get_random_string(length):
-    letters = string.ascii_lowercase
-    return "".join(random.choice(letters) for i in range(length))
-
-
-@pytest.fixture
-def client() -> XataClient:
-    return XataClient()
-
-
-@pytest.fixture
-def demo_db(client: XataClient) -> string:
-    db_name = f"sdk-py-e2e-test-{get_random_string(6)}"
-    create_demo_db(client, db_name)
-    client.set_db_and_branch_names(db_name, "main")
-    yield db_name
-    delete_db(client, db_name)
-
-
-@pytest.fixture
-def posts() -> list[str]:
-    return [
-        {
-            "title": "Hello world",
-            "labels": ["hello", "world"],
-            "slug": "hello-world",
-            "text": "This is a test post",
-        },
-        {
-            "title": "HeLLo universe",
-            "labels": ["hello", "universe"],
-            "slug": "hello-universe",
-            "text": "hello, is it me you're looking for?",
-        },
-        {
-            "title": "HELlO internet",
-            "labels": ["hello", "internet"],
-            "slug": "hello-internet",
-            "text": "I like to eat apples and bananas",
-        },
-    ]
-
-
-def _wait_until_records_are_indexed(table: str):
-    """
-    Wait for the records to be index in order to able to search them
-    """
-    # TODO remove in favour of wait loop with aggs
-    # when aggs are available
-    time.sleep(10)
 
 
 # ------------------------------------------------------- #
@@ -301,136 +205,4 @@ def test_delete_record(client: XataClient, demo_db: string):
 
     with pytest.raises(RecordNotFoundException) as exc:
         client.delete_record("Posts", recId)
-    assert exc is not None
-
-
-# ------------------------------------------------------- #
-#
-# Search
-#
-# ------------------------------------------------------- #
-def test_search_simple(client: XataClient, demo_db: string, posts: list[str]):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search("hello")
-    assert "records" in result
-    assert len(result["records"]) == len(posts)
-
-    result = client.search("apples")
-    assert "records" in result
-    assert len(result["records"]) == 1
-
-
-def test_search_with_params(client: XataClient, demo_db: string, posts: list[str]):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search(
-        "hello",
-        {
-            "fuzziness": 1,
-            "prefix": "phrase",
-        },
-    )
-    assert "records" in result
-    assert len(result["records"]) == len(posts)
-
-    result = client.search(
-        "apples and bananas",
-        {
-            "fuzziness": 0,
-            "prefix": "phrase",
-        },
-    )
-    assert "records" in result
-    assert len(result["records"]) == 1
-
-
-def test_search_with_no_hits(client: XataClient, demo_db: string, posts: list[str]):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search("12345")
-    assert "records" in result
-    assert len(result["records"]) == 0
-
-
-def test_search_errorcases(client: XataClient, demo_db: string, posts: list[str]):
-    with pytest.raises(BadRequestException) as exc:
-        client.search("invalid", {"i-am": "invalid"})
-    assert exc is not None
-
-
-# ------------------------------------------------------- #
-#
-# Search Table
-#
-# ------------------------------------------------------- #
-def test_search_table_simple(client: XataClient, demo_db: string, posts: list[str]):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search_table("Posts", "hello")
-    assert "records" in result
-    assert len(result["records"]) == len(posts)
-
-    result = client.search_table("Posts", "apples")
-    assert "records" in result
-    assert len(result["records"]) == 1
-
-
-def test_search_table_with_params(
-    client: XataClient, demo_db: string, posts: list[str]
-):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search_table(
-        "Posts",
-        "hello",
-        {
-            "fuzziness": 1,
-            "prefix": "phrase",
-        },
-    )
-    assert "records" in result
-    assert len(result["records"]) == len(posts)
-
-    result = client.search_table(
-        "Posts",
-        "apples and bananas",
-        {
-            "fuzziness": 0,
-            "prefix": "phrase",
-        },
-    )
-    assert "records" in result
-    assert len(result["records"]) == 1
-
-
-def test_search_table_with_no_hits(
-    client: XataClient, demo_db: string, posts: list[str]
-):
-    for post in posts:
-        client.create("Posts", record=post)
-    _wait_until_records_are_indexed("Posts")
-
-    result = client.search_table("Posts", "watermelon")
-    assert "records" in result
-    assert len(result["records"]) == 0
-
-
-def test_search_table_errorcases(client: XataClient, demo_db: string, posts: list[str]):
-    result = client.search_table("MissingTable", "hello")
-    assert "message" in result
-    assert result["message"] == f"table [{demo_db}:main/MissingTable] not found"
-
-    with pytest.raises(BadRequestException) as exc:
-        client.search_table("Posts", "invalid", {"i-am": "invalid"})
     assert exc is not None
