+++ tests/integration-tests/utils.py
@@ -1,0 +1,28 @@
+#
+# Licensed to Xatabase, Inc under one or more contributor
+# license agreements. See the NOTICE file distributed with
+# this work for additional information regarding copyright
+# ownership. Xatabase, Inc licenses this file to you under the
+# Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You
+# may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+import time
+
+def wait_until_records_are_indexed(table: str):
+    """
+    Wait for the records to be index in order to able to search them
+    """
+    # TODO remove in favour of wait loop with aggs
+    # when aggs are available
+    time.sleep(10)
